+++ packages/app-staking/src/Account/index.tsx
@@ -5,16 +5,17 @@
 import { DerivedBalancesMap } from '@polkadot/api-derive/types';
 import { I18nProps } from '@polkadot/ui-app/types';
 import { ApiProps } from '@polkadot/ui-api/types';
+import { Nominators } from '../types';
 
 import React from 'react';
-import { AccountId, Balance, Option, StakingLedger, ValidatorPrefs } from '@polkadot/types';
+import { AccountId, Balance, Exposure, Option, StakingLedger, ValidatorPrefs } from '@polkadot/types';
 import { AddressMini, AddressSummary, Button, TxButton } from '@polkadot/ui-app';
 import { withCalls } from '@polkadot/ui-api';
 
 import Bonding from './Bonding';
-import Controller from './Controller';
-import Staking from './Staking';
 import Nominating from './Nominating';
+import SessionKey from './SessionKey';
+import Validating from './Validating';
 import translate from '../translate';
 
 type Props = ApiProps & I18nProps & {
@@ -25,49 +26,47 @@ type Props = ApiProps & I18nProps & {
   session_nextKeyFor?: Option<AccountId>,
   staking_bonded?: Option<AccountId>,
   staking_ledger?: Option<StakingLedger>,
-  staking_nominating?: Option<AccountId>,
-  staking_nominatorsFor?: Array<string>,
-  staking_validators?: ValidatorPrefs,
+  staking_stakers?: Exposure,
+  staking_validators?: [ValidatorPrefs],
   intentions: Array<string>,
+  nominators: Nominators,
   isValidator: boolean,
   validators: Array<string>
 };
 
 type State = {
   isBondingOpen: boolean,
-  isControllerOpen: boolean,
   isNominateOpen: boolean,
-  isStakingOpen: boolean,
+  isSessionKeyOpen: boolean,
+  isValidatingOpen: boolean,
   isUnbondOpen: boolean,
   bondedId: string | null,
-  controllerId: string | null,
-  nomineeId: string | null,
+  sessionId: string | null,
   stashId: string | null
 };
 
 class Account extends React.PureComponent<Props, State> {
   state: State = {
     isBondingOpen: false,
-    isControllerOpen: false,
+    isSessionKeyOpen: false,
     isNominateOpen: false,
-    isStakingOpen: false,
+    isValidatingOpen: false,
     isUnbondOpen: false,
     bondedId: null,
-    controllerId: null,
-    nomineeId: null,
+    sessionId: null,
     stashId: null
   };
 
-  static getDerivedStateFromProps ({ session_nextKeyFor, staking_bonded, staking_ledger, staking_nominating }: Props): Partial<State> {
+  static getDerivedStateFromProps ({ session_nextKeyFor, staking_bonded, staking_ledger }: Props): Partial<State> {
+
+    // console.error('staking_stakers', JSON.stringify(staking_stakers));
+
     return {
       bondedId: staking_bonded && staking_bonded.isSome
         ? staking_bonded.unwrap().toString()
         : null,
-      controllerId: session_nextKeyFor && session_nextKeyFor.isSome
+      sessionId: session_nextKeyFor && session_nextKeyFor.isSome
         ? session_nextKeyFor.unwrap().toString()
-        : null,
-      nomineeId: staking_nominating && staking_nominating.isSome
-        ? staking_nominating.unwrap().toString()
         : null,
       stashId: staking_ledger && staking_ledger.isSome
         ? staking_ledger.unwrap().stash.toString()
@@ -81,9 +80,9 @@ class Account extends React.PureComponent<Props, State> {
     return (
       <article className='staking--Account'>
         {this.renderBonding()}
-        {this.renderController()}
         {this.renderNominating()}
-        {this.renderStaking()}
+        {this.renderSessionKey()}
+        {this.renderValidating()}
         <AddressSummary
           balance={balanceArray(accountId)}
           name={name}
@@ -92,6 +91,8 @@ class Account extends React.PureComponent<Props, State> {
         >
           <div className='staking--Account-expand'>
             {this.renderButtons()}
+            {this.renderBondedId()}
+            {this.renderStashId()}
             {this.renderNominee()}
             {this.renderNominators()}
           </div>
@@ -102,51 +103,46 @@ class Account extends React.PureComponent<Props, State> {
 
   private renderBonding () {
     const { accountId } = this.props;
-    const { controllerId, isBondingOpen } = this.state;
-
-    if (!controllerId) {
-      return null;
-    }
+    const { bondedId, isBondingOpen } = this.state;
 
     return (
       <Bonding
         accountId={accountId}
-        controllerId={controllerId}
+        bondedId={bondedId}
         isOpen={isBondingOpen}
         onClose={this.toggleBonding}
       />
     );
   }
 
-  private renderController () {
-    const { accountId } = this.props;
-    const { controllerId, isControllerOpen } = this.state;
-
-    return (
-      <Controller
-        accountId={accountId}
-        controllerId={controllerId}
-        isOpen={isControllerOpen}
-        onClose={this.toggleController}
-      />
-    );
-  }
-
-  private renderStaking () {
+  private renderValidating () {
     const { accountId, staking_validators } = this.props;
-    const { isStakingOpen, stashId } = this.state;
+    const { isValidatingOpen, stashId } = this.state;
 
-    if (!staking_validators || !isStakingOpen || !stashId) {
+    if (!staking_validators || !isValidatingOpen || !stashId) {
       return null;
     }
 
     return (
-      <Staking
+      <Validating
         accountId={accountId}
         isOpen
-        onClose={this.toggleStaking}
-        preferences={staking_validators}
+        onClose={this.toggleValidating}
+        preferences={staking_validators[0]}
         stashId={stashId}
+      />
+    );
+  }
+
+  private renderSessionKey () {
+    const { accountId } = this.props;
+    const { isSessionKeyOpen } = this.state;
+
+    return (
+      <SessionKey
+        accountId={accountId}
+        isOpen={isSessionKeyOpen}
+        onClose={this.toggleSessionKey}
       />
     );
   }
@@ -169,46 +165,98 @@ class Account extends React.PureComponent<Props, State> {
   //   );
   // }
 
-  private renderNominee () {
-    const { balanceArray } = this.props;
-    const { nomineeId } = this.state;
+  private getNominees () {
+    const { nominators } = this.props;
+    const { stashId } = this.state;
 
-    if (!nomineeId) {
+    return stashId ? nominators[stashId] : null;
+  }
+
+  private renderNominee () {
+    const nominees = this.getNominees();
+
+    if (!nominees || !nominees.length) {
       return null;
     }
 
     return (
-      <AddressMini
-        balance={balanceArray(nomineeId)}
-        value={nomineeId}
-        withBalance
-      />
+      <div className='staking--Account-detail'>
+        <label className='staking--label'>nominating</label>
+        {
+          nominees.map((nomineeId, index) => (
+            <AddressMini
+              key={index}
+              value={nomineeId}
+              withBalance
+            />
+          ))
+        }
+      </div>
     );
   }
 
   private renderNominators () {
-    const { staking_nominatorsFor } = this.props;
+    // const { staking_nominatorsFor } = this.props;
 
-    if (!staking_nominatorsFor) {
+    // if (!staking_nominatorsFor) {
+    //   return null;
+    // }
+
+    // return (
+    //   <div className='staking--Account-detail'>
+    //     <label className='staking--label'>nominators</label>
+    //     <div className='ui--Nominators'>
+    //       {staking_nominatorsFor.map((nominator) => (
+    //         <AddressMini
+    //           isPadded={false}
+    //           key={nominator}
+    //           value={nominator}
+    //           withBalance
+    //         />
+    //       ))}
+    //     </div>
+    //   </div>
+    // );
+  }
+
+  private renderBondedId () {
+    const { bondedId } = this.state;
+
+    if (!bondedId) {
       return null;
     }
 
     return (
-      <div className='ui--Nominators'>
-        {staking_nominatorsFor.map((nominator) => (
-          <AddressMini
-            isPadded={false}
-            key={nominator}
-            value={nominator}
-            withBalance
-          />
-        ))}
+      <div className='staking--Account-detail'>
+        <label className='staking--label'>controller account</label>
+        <AddressMini
+          value={bondedId}
+          withBalance
+        />
+      </div>
+    );
+  }
+
+  private renderStashId () {
+    const { stashId } = this.state;
+
+    if (!stashId) {
+      return null;
+    }
+
+    return (
+      <div className='staking--Account-detail'>
+        <label className='staking--label'>stash account</label>
+        <AddressMini
+          value={stashId}
+          withBalance
+        />
       </div>
     );
   }
 
   private renderNominating () {
-    const { accountId, intentions, validators } = this.props;
+    const { accountId, intentions } = this.props;
     const { isNominateOpen, stashId } = this.state;
 
     if (!stashId) {
@@ -222,29 +270,17 @@ class Account extends React.PureComponent<Props, State> {
         onClose={this.toggleNominate}
         intentions={intentions}
         stashId={stashId}
-        validators={validators}
       />
     );
   }
 
   private renderButtons () {
-    const { accountId, t } = this.props;
-    const { controllerId, stashId } = this.state;
+    const { accountId, intentions, t } = this.props;
+    const { sessionId, stashId } = this.state;
     const buttons = [];
 
     if (!stashId) {
-      buttons.push(
-        <Button
-          isPrimary
-          key='controller'
-          onClick={this.toggleController}
-          label={t('Controller')}
-        />
-      );
-
-      // only display bonding if we already have a controller
-      if (controllerId) {
-        buttons.push(<Button.Or key='bond.or' />);
+      if (sessionId) {
         buttons.push(
           <Button
             isPrimary
@@ -253,44 +289,50 @@ class Account extends React.PureComponent<Props, State> {
             label={t('Bond')}
           />
         );
+      } else {
+        buttons.push(
+          <Button
+            isPrimary
+            key='session'
+            onClick={this.toggleSessionKey}
+            label={t('Set Session Key')}
+          />
+        );
       }
     } else {
-      // buttons.push(
-      //   <Button
-      //     isNegative
-      //     label={t('Unbond')}
-      //     key='unbond'
-      //     onClick={this.toggleUnbond}
-      //   />
-      // );
-      // buttons.push(<Button.Or key='stake.or' />);
-      buttons.push(
-        <Button
-          isPrimary
-          key='stake'
-          onClick={this.toggleStaking}
-          label={t('Stake')}
-        />
-      );
-      buttons.push(<Button.Or key='nominate.or' />);
-      buttons.push(
-        <Button
-          isPrimary
-          key='nominate'
-          onClick={this.toggleNominate}
-          label={t('Nominate')}
-        />
-      );
-      buttons.push(<Button.Or key='stop.or' />);
-      buttons.push(
-        <TxButton
-          accountId={accountId}
-          isNegative
-          label={t('Stop')}
-          key='stop'
-          tx='staking.chill'
-        />
-      );
+      const nominees = this.getNominees();
+      const isNominating = nominees && nominees.length;
+      const isValidating = intentions.indexOf(accountId) !== -1;
+
+      if (isValidating || isNominating) {
+        buttons.push(
+          <TxButton
+            accountId={accountId}
+            isNegative
+            label={isValidating ? t('Stop Validating') : t('Stop Nominating')}
+            key='stop'
+            tx='staking.chill'
+          />
+        );
+      } else {
+        buttons.push(
+          <Button
+            isPrimary
+            key='validate'
+            onClick={this.toggleValidating}
+            label={t('Validate')}
+          />
+        );
+        buttons.push(<Button.Or key='nominate.or' />);
+        buttons.push(
+          <Button
+            isPrimary
+            key='nominate'
+            onClick={this.toggleNominate}
+            label={t('Nominate')}
+          />
+        );
+      }
     }
 
     return (
@@ -306,29 +348,23 @@ class Account extends React.PureComponent<Props, State> {
     }));
   }
 
-  private toggleController = () => {
-    this.setState(({ isControllerOpen }) => ({
-      isControllerOpen: !isControllerOpen
-    }));
-  }
-
   private toggleNominate = () => {
     this.setState(({ isNominateOpen }) => ({
       isNominateOpen: !isNominateOpen
     }));
   }
 
-  private toggleStaking = () => {
-    this.setState(({ isStakingOpen }) => ({
-      isStakingOpen: !isStakingOpen
+  private toggleSessionKey = () => {
+    this.setState(({ isSessionKeyOpen }) => ({
+      isSessionKeyOpen: !isSessionKeyOpen
     }));
   }
 
-  // private toggleUnbond = () => {
-  //   this.setState(({ isUnbondOpen }) => ({
-  //     isUnbondOpen: !isUnbondOpen
-  //   }));
-  // }
+  private toggleValidating = () => {
+    this.setState(({ isValidatingOpen }) => ({
+      isValidatingOpen: !isValidatingOpen
+    }));
+  }
 }
 
 export default translate(
@@ -336,8 +372,7 @@ export default translate(
     ['query.session.nextKeyFor', { paramName: 'accountId' }],
     ['query.staking.bonded', { paramName: 'accountId' }],
     ['query.staking.ledger', { paramName: 'accountId' }],
-    ['query.staking.nominatorsFor', { paramName: 'accountId' }],
-    ['query.staking.nominating', { paramName: 'accountId' }],
+    ['query.staking.stakers', { paramName: 'accountId' }],
     ['query.staking.validators', { paramName: 'accountId' }]
   )(Account)
 );