+++ packages/app-staking/src/index.tsx
@@ -5,13 +5,15 @@
 import { DerivedBalancesMap } from '@polkadot/api-derive/types';
 import { AppProps, I18nProps } from '@polkadot/ui-app/types';
 import { ApiProps } from '@polkadot/ui-api/types';
-import { ComponentProps } from './types';
+import { SubjectInfo } from '@polkadot/ui-keyring/observable/types';
+import { ComponentProps, Nominators } from './types';
 
 import React from 'react';
 import { Route, Switch } from 'react-router';
-import { AccountId, Balance } from '@polkadot/types';
+import { AccountId, Balance, Option } from '@polkadot/types';
 import Tabs, { TabItem } from '@polkadot/ui-app/Tabs';
-import { withCalls, withMulti } from '@polkadot/ui-api';
+import { withCalls, withMulti, withObservable } from '@polkadot/ui-api';
+import accountObservable from '@polkadot/ui-keyring/observable/accounts';
 
 import './index.css';
 
@@ -20,13 +22,16 @@ import Overview from './Overview';
 import translate from './translate';
 
 type Props = AppProps & ApiProps & I18nProps & {
+  allAccounts?: SubjectInfo,
   balances?: DerivedBalancesMap,
-  intentions?: Array<AccountId>,
-  session_validators?: Array<AccountId>
+  session_validators?: Array<AccountId>,
+  staking_controllers?: [Array<AccountId>, Array<Option<AccountId>>],
+  staking_nominators?: [Array<AccountId>, Array<Array<AccountId>>]
 };
 
 type State = {
   intentions: Array<string>,
+  nominators: Nominators,
   tabs: Array<TabItem>,
   validators: Array<string>
 };
@@ -41,6 +46,7 @@ class App extends React.PureComponent<Props, State> {
 
     this.state = {
       intentions: [],
+      nominators: {},
       tabs: [
         {
           name: 'overview',
@@ -55,26 +61,37 @@ class App extends React.PureComponent<Props, State> {
     };
   }
 
-  static getDerivedStateFromProps ({ session_validators, intentions }: Props): State {
+  static getDerivedStateFromProps ({ staking_controllers = [[], []], session_validators = [], staking_nominators = [[], []] }: Props): State {
     return {
-      intentions: (intentions || []).map((accountId) =>
-        accountId.toString()
+      intentions: staking_controllers[1].filter((optId) => optId.isSome).map((accountId) =>
+        accountId.unwrap().toString()
       ),
-      validators: (session_validators || []).map((authorityId) =>
+      nominators: staking_nominators[0].reduce((result, accountId, index) => {
+        result[accountId.toString()] = staking_nominators[1][index].map((accountId) =>
+          accountId.toString()
+        );
+
+        return result;
+      }, {} as Nominators),
+      validators: session_validators.map((authorityId) =>
         authorityId.toString()
       )
     } as State;
   }
 
   render () {
+    const { allAccounts, basePath } = this.props;
     const { tabs } = this.state;
-    const { basePath } = this.props;
+    const hidden = !allAccounts || Object.keys(allAccounts).length === 0
+      ? ['actions']
+      : [];
 
     return (
       <main className='staking--App'>
         <header>
           <Tabs
             basePath={basePath}
+            hidden={hidden}
             items={tabs}
           />
         </header>
@@ -88,7 +105,7 @@ class App extends React.PureComponent<Props, State> {
 
   private renderComponent (Component: React.ComponentType<ComponentProps>) {
     return (): React.ReactNode => {
-      const { intentions, validators } = this.state;
+      const { intentions, nominators, validators } = this.state;
       const { balances = {} } = this.props;
 
       return (
@@ -96,6 +113,7 @@ class App extends React.PureComponent<Props, State> {
           balances={balances}
           balanceArray={this.balanceArray}
           intentions={intentions}
+          nominators={nominators}
           validators={validators}
         />
       );
@@ -124,8 +142,10 @@ export default withMulti(
   App,
   translate,
   withCalls<Props>(
+    'derive.staking.controllers',
     'query.session.validators',
-    ['query.staking.intentions', { propName: 'intentions' }],
+    'query.staking.nominators',
     ['derive.staking.intentionsBalances', { propName: 'balances' }]
-  )
+  ),
+  withObservable(accountObservable.subject, { propName: 'allAccounts' })
 );